pyproject.toml
[project]
name = "ghost-mcp"
version = "0.1.0"
description = "Ghost blog integration MCP server"
authors = [
    { name = "Fanyang Meng" }
]
dependencies = [
    "httpx",
    "pyjwt",
    "mcp[cli]>=1.2.1"
]
requires-python = ">=3.12"
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ghost_mcp"]

[tool.hatch.metadata]
allow-direct-references = true
